requirements.txt
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
importlib-metadata==5.1.0
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2==2.9.5
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
vine==5.0.0
wcwidth==0.2.5
yoyo-migrations==8.1.0
zipp==3.11.0
